added admin panel for users

+++ app/Http/Controllers/UserController.php
@@ -3,8 +3,12 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Http\Requests\UserRequest;
 use App\Repositories;
+use App\Repositories\UserRepository;
 use App\Http\Requests;
+use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -22,7 +26,14 @@ class UserController extends Controller
      */
     public function index()
     {
-        //
+        $route = Route::current();
+        if($route->role) {
+            $users = $this->userRepository->getUsersSortBy($route->role);
+            }
+        else {
+            $users = $this->userRepository->getUsersSortBy();
+        }
+        return view('user/index', compact('users'));
     }
 
     /**
@@ -32,7 +43,7 @@ class UserController extends Controller
      */
     public function create()
     {
-        //
+        return view('user/create');
     }
 
     /**
@@ -44,6 +55,7 @@ class UserController extends Controller
     public function store(UserRequest $request)
     {
         $this->userRepository->store($request->all());
+        return redirect('/user');
     }
 
     /**
@@ -66,6 +78,7 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = $this->userRepository->getById($id);
+        return view('user/edit', compact('user'));
     }
 
     /**
@@ -78,6 +91,7 @@ class UserController extends Controller
     public function update(UserRequest $request, $id)
     {
         $this->userRepository->update($id, $request->all());
+        return redirect('user');
     }
 
     /**
@@ -88,6 +102,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $this->userRepository->delete($id);
+        if(Auth::user()->id != $id)
+            $this->userRepository->delete($id);
+        return back();
+    }
+
+    public function changeSeen(Request $request) {
+        $this->userRepository->changeUserSeen($request['id']);
+        return \response()->json($request['id']);
+    }
+
+    public function changeAdmin(Request $request) {
+        $this->userRepository->changeUserAdmin($request['id']);
+        return \response()->json($request['id']);
     }
 }
